Fixed bug in config.secret_key code
The <% tags are not needed, and would cause a bug:

/.rvm/gems/ruby-2.0.0-p247/gems/activesupport-3.2.13/lib/active_support/dependencies.rb:245:in `load': /Users/Magne/Workspace/bloggery/config/initializers/devise.rb:9: syntax error, unexpected '<' (SyntaxError)
<% if rails_4? -%>
 ^
/projectname/config/initializers/devise.rb:9: syntax error, unexpected tFID, expecting keyword_end
<% if rails_4? -%>
              ^

Also, rails_4 is in the  `devise/lib/generators/devise/install_generator.rb` file, which is not automatically included in a project when using the latest devise gem. So people will probably not have this file. Therefore it's better to use `Rails::VERSION::MAJOR == 4` directly instead.

PS: I'm on rails 3.2.13 and devise 3.4.1.