Allow redirect whitelist & block tld redirects
It's currently possible for an attacker to initiate an open redirect
by passing `?redirect=.evil.com` param, which will redirect to the new
top level domain.

For example:
https://example.com/signin?redirect=.evil.com

Upon sign in will redirect to:
https://example.com.evil.com

This locks down what can be redirected to only paths beginning with
/, ?, and #.

It also enables configuring a whitelist of domains that can be
redirected to if you would like.