Allow overriding the source of the password_reset_token
Changes no behavior by default, but this allows for pulling the token
out of the session for example.

The motivation behind this is to stop leaking the token to third parties
when tracking JavaScript is present on the page (e.g. Google Analytics).

We're overriding `edit` to set the token in the session and then redirect,
and this change would enable us to be slightly less coupled to Devise's
internals.