chore: add expiry note for old rack/rails session bug fix
Seven year ago rails `session.keys` could be empty if the session was
not loaded yet.

To prevent an error the removed code was introduced
https://github.com/heartcombo/devise/issues/2660

Since then rails changed the behaviour and makes sure that the session
is loaded before someone wants to access any session keys
https://github.com/rails/rails/commit/3498aacbbebb41e529b6755f4ccfdfbb84c28830

Which means the `session.empty?` is not needed anymore once Rails 5.2+
and upwards only supported.