Add a setting to keep the unlock token after unlocking a resource
By default, Devise clears the locked_at, failed_attempts and
unlock_token when unlocking a resource. Even though the unlock token
doesn't have an impact on whether a resource is locked or not, such a
change can impact users who depend on this column being not empty. To
address this problem, I'm introducing the new behavior of not cleaning
the unlock token as a opt-in setting.