Keep unlock token after unlocking when configured
The path GET /resource/unlock?unlock_token=TOKEN is not currently
idempotent, which brings some problems when the page is hit by an
antivirus, prefetch or anything other than the user. For example,
Microsoft SafeLinks technology visits URLs sent in emails to make sure
that the URL doesn't have malware.

This change keeps, when configured, the unlock token after the resource
has been unlocked. If the user happens to click on the unlock link after
it has already been visited, it will behave as if this was the first
visit, even though the resource is technically unlocked since the first
visit. This makes the GET request idempotent and doesn't break the
current behavior.

This may add a side effect on database queries that check how many
resources have unlock_token not null to count how many resources are
unlocked. To address this side effect, the new behavior is turned off by
default.

The [first version] of the Lockable module logged the user in, but this
is not true since #1493 was merged. When a resource is locked again,
another unlock token is created and overwrites the older one. Keeping
the most recent unlock token, then, doesn't represent a security issue.

[first version]:
https://github.com/heartcombo/devise/commit/d2fa737aa094f2e4808d9d4388b20d7081861511