Keep the number of failed attempts on unlock revisits
We want to avoid malicious actors who possess the last unlock token of a
user to have a way to be able to reset the number of failed attempts
just by visiting the unlock URL.

By early returning `true` on the `unlock_access!` method if the
`locked_at` is blank - i.e. the resource is not locked -, Devise can
behave as if the lock was removed without resetting the failed login
attempts.