Add support for namespaced Devise models in views path
This commit modifies the `add_devise_routes` method to handle namespaced Devise models correctly. When the `class_name` includes a namespace (i.e., contains "::"), the namespace is used as the `as` option value for the `devise_for` route helper. This ensures that Devise looks for views in the correct directory corresponding to the namespace of the model.