Replace I18n.available_locales with I18n.backend.eager_load!
Currently I18n.available_locales is called from the tests to
ensure that the I18n backend is initialized before the tests
run. This, however, is not only unintuitive to initialize the
backend, but also unreliable because if the available locales
are cached, calling I18n.available_locales won't call further
into the backend to initialize it.

This is used in combination with I18n.reload! which leaves the
backend in uninitialized state but it doesn't clear the global
I18n.available_locales, i.e. the available locales are still
cached. Consequently, I18n.available_locales won't do what's
expected of it - namely initializing the backend by preloading
the translations - and tests relying on custom translations
may break when the I18n backend initializes itself while
looking up a translation key. During this init phase the I18n
backend may override translations defined in the tests by
translations loaded from the locale YML files.

I18n.backend.eager_load! is independent of the cached available
locales. If the backend is already initialized, it returns
early, otherwise it loads the translations from the YML files.